pyproject.toml
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project]
name = "api_pgd_client"
version = "0.1.0"
description = "Biblioteca para simplificar as chamadas à API do Programa de Gestão e Desempenho do Ministério da Gestão (GovBR) para sistemas em Python, que manipula dados do teletrabalho de servidores da administração pública."
readme = "README.rst"
authors = [
  {name = "Ivan Medeiros Jr", email = "ivsmjunior@gmail.com"}
]
maintainers = [
  {name = "Ivan Medeiros Jr", email = "ivsmjunior@gmail.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities"
]
keywords = [ "pgd", "programa-de-gestao", "govbr", "api", "python", "teletrabalho" ]
license = {text = "MIT license"}
requires-python = ">=3.9"
dependencies = [
    "requests (>=2.31.0,<3.0.0)",
    "python-decouple (>=3.8,<4.0)",
]

[project.optional-dependencies]
dev = [
    "mypy",  # linting
    "pytest",  # testing
    "ruff"  # linting
]

[project.urls]

"Bug Tracker" = "https://github.com/ivanjun10r/api_pgd_client/issues"
changelog = "https://github.com/ivanjun10r/api_pgd_client/blob/master/changelog.md"
homepage = "https://github.com/ivanjun10r/api_pgd_client"
repository = "https://github.com/ivanjun10r/api_pgd_client"


# Poetry
# ------

[tool.poetry]
packages = [
    { include = "api_pgd_client", from = "src" }
]

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"
bump2version = "^1.0.1"
watchdog = "^6.0.0"
sphinx = "7.2.6"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.2"
pytest-cov = "^4.1.0"
pytest-reverse = "^1.8.0"


# Mypy
# ----
[tool.mypy]
files = "."
# Use strict defaults
strict = true
warn_unreachable = true
warn_no_return = true
ignore_missing_imports = true
exclude = "tests/"
python_version = "3.9"

# Isort
# -----
[tool.isort]
profile = "black"
line_length = 80
known_first_party = "pgd_api"
multi_line_output = 3
include_trailing_comma = true
use_parentheses = true


# Pytest
# ------
[tool.pytest.ini_options]
pythonpath = "src"
addopts = "-v --cov=api_pgd_client"
